MXS-2313: Use 64-bit integers to store rank
Although the default value is the maximum value of a signed 32-bit
integer, the value is stored as a 64-bit integer. The integer type
conversion functions return 64-bit values so storing it as one makes
sense.

Currently values higher than the default are allowed but the accepted
range of input should be restricted in the future.

+++ include/maxscale/server.hh
@@ -282,7 +282,7 @@ public:
      *
      * @return The server rank
      */
-    virtual int rank() const = 0;
+    virtual int64_t rank() const = 0;
 
     /*
      * Update server address. TODO: Move this to internal class once blr is gone.
