+++ server/core/internal/server.hh
@@ -62,7 +62,7 @@ public:
         m_settings.persistmaxtime = persistmaxtime;
     }
 
-    void set_rank(int rank)
+    void set_rank(int64_t rank)
     {
         m_settings.rank = rank;
     }
@@ -122,7 +122,7 @@ public:
         return m_settings.authenticator;
     }
 
-    int rank() const override
+    int64_t rank() const override
     {
         return m_settings.rank;
     }
@@ -341,7 +341,7 @@ private:
         long persistpoolmax = 0;    /**< Maximum size of persistent connections pool */
         long persistmaxtime = 0;    /**< Maximum number of seconds connection can live */
 
-        int rank;   /*< The ranking of this server, used to prioritize certain servers over others */
+        int64_t rank;   /*< The ranking of this server, used to prioritize certain servers over others */
 
         /** Disk space thresholds. Can be queried from modules at any time so access must be protected
          *  by mutex. */
